adding updated code for the STT file, created env file and dialogapi folder
