+++ Develop/config/config.json
@@ -1,4 +1,11 @@
 {
+  "martin": {
+    "username": "root",
+    "password": "Popo1234",
+    "database": "pet_speak_health_db",
+    "host": "127.0.0.1",
+    "dialect": "mysql"
+  },
   "development": {
     "username": "root",
     "password": "rootpw",
